+++ libcnb/src/build.rs
@@ -1,3 +1,5 @@
+//! Provides build phase specific types and helpers.
+
 use std::{fs, path::PathBuf};
 
 use serde::de::DeserializeOwned;
@@ -27,7 +29,7 @@ pub struct BuildContext<B: Buildpack + ?Sized> {
 /// Describes the result of the build phase.
 ///
 /// In contrast to `DetectResult`, it always signals a successful build. To fail the build phase,
-/// return a failed [`libcnb::Result`] from the build function.
+/// return a failed [`crate::Result`] from the build function.
 ///
 /// It contains build phase output such as launch and/or store metadata which will be subsequently
 /// handled by libcnb.