+++ libcnb/src/generic.rs
@@ -1,6 +1,9 @@
+//! Generic implementations for some libcnb types.
+
 use std::path::Path;
 
-use crate::platform::{Platform, PlatformEnv};
+use crate::platform::Platform;
+use crate::{read_platform_env, Env};
 use std::fmt::{Debug, Display, Formatter};
 
 /// Generic TOML metadata.
@@ -20,17 +23,21 @@ impl Display for GenericError {
 
 /// A generic platform that only provides access to environment variables.
 pub struct GenericPlatform {
-    env: PlatformEnv,
+    env: Env,
+}
+
+impl GenericPlatform {
+    pub fn new(env: Env) -> Self {
+        Self { env }
+    }
 }
 
 impl Platform for GenericPlatform {
-    fn env(&self) -> &PlatformEnv {
+    fn env(&self) -> &Env {
         &self.env
     }
 
     fn from_path(platform_dir: impl AsRef<Path>) -> std::io::Result<Self> {
-        Ok(Self {
-            env: PlatformEnv::from_path(platform_dir)?,
-        })
+        read_platform_env(platform_dir.as_ref()).map(|env| GenericPlatform { env })
     }
 }